implement general interpolation environment
Closes #3

Three interpolation functions are defined by default:

* \EaseInterpolate -- for easing animations
* \IdentityInterpolate -- linear interpolation
* \ReverseInterpolate -- reverse the animation

Interpolations work with the `interpolate` environment that expects one
argument which is the name of the interpolation function.

Interpolation functions are maps [0, 1] -> [0, 1] which are used to
redefine the \fprogress macro. As such, they are agnostic to the the
exact frame number and care only about how far along we are in the
animation.

Interpolations cannot be nested. This will cause an infinite loop in
LaTeX. I believe this can be solved by introducing a counter to generate
unique intermediate \oldfprogress macros.